most_damage battle_bot
