+++ showdown/battle_bots/safest/main.py
@@ -1,7 +1,39 @@
 from showdown.battle import Battle
 
-from .helpers import pick_safest_move_from_battles
-from .helpers import format_decision
+from ..helpers import format_decision
+
+from showdown.engine.objects import StateMutator
+from showdown.engine.select_best_move import pick_safest
+from showdown.engine.select_best_move import get_payoff_matrix
+
+from config import logger
+
+
+def prefix_opponent_move(score_lookup, prefix):
+    new_score_lookup = dict()
+    for k, v in score_lookup.items():
+        bot_move, opponent_move = k
+        new_opponent_move = "{}_{}".format(opponent_move, prefix)
+        new_score_lookup[(bot_move, new_opponent_move)] = v
+
+    return new_score_lookup
+
+
+def pick_safest_move_from_battles(battles):
+    all_scores = dict()
+    for i, b in enumerate(battles):
+        state = b.create_state()
+        mutator = StateMutator(state)
+        user_options, opponent_options = b.get_all_options()
+        logger.debug("Attempting to find best move from: {}".format(mutator.state))
+        scores = get_payoff_matrix(mutator, user_options, opponent_options, prune=True)
+        prefixed_scores = prefix_opponent_move(scores, str(i))
+        all_scores = {**all_scores, **prefixed_scores}
+
+    decision, payoff = pick_safest(all_scores)
+    bot_choice = decision[0]
+    logger.debug("Safest: {}, {}".format(bot_choice, payoff))
+    return bot_choice
 
 
 class BattleBot(Battle):
